feat: 왜 왜 안되는거야

+++ philo/include/philo.h
@@ -6,7 +6,7 @@
 /*   By: hyeongki <hyeongki@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/23 19:43:19 by hyeongki          #+#    #+#             */
-/*   Updated: 2022/11/12 16:41:10 by hyeongki         ###   ########.fr       */
+/*   Updated: 2022/11/13 18:17:25 by hyeongki         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -52,7 +52,7 @@ typedef struct s_philo
 	int				number;
 	int				fork[2];
 	int				eat_cnt;
-	int				eat_start_time;
+	uint64_t		eat_start_time;
 	int				status;
 	struct s_table	*table;
 	pthread_mutex_t	status_mutex;
@@ -61,10 +61,10 @@ typedef struct s_philo
 typedef struct s_table
 {
 	int				number_of_philo;
-	int				time_to_die;
-	int				time_to_eat;
-	int				time_to_sleep;
-	int				start_time;
+	uint64_t		time_to_die;
+	uint64_t		time_to_eat;
+	uint64_t		time_to_sleep;
+	uint64_t		start_time;
 	int				number_of_eat;
 	int				noe_flag;
 	int				finish;
@@ -77,10 +77,10 @@ typedef struct s_table
 int		init_table(int argc, char **argv, t_table *table);
 
 /* util.c */
-int		ft_strlen(char *str);
-int		ft_atoi(char *str);
-int		get_current_time(void);
-int		is_dead(t_table *table, t_philo *philo);
+int			ft_strlen(char *str);
+int			ft_atoi(char *str);
+uint64_t	get_current_time(void);
+int			is_dead(t_table *table, t_philo *philo);
 
 /* error.c */
 int		put_error(int err);
@@ -89,7 +89,7 @@ int		argment_check(t_table *table);
 
 /* simulation */
 int		simulation(t_table *table);
-int		philo_usleep(t_philo *philo, int time);
+int		philo_usleep(t_philo *philo, uint64_t time);
 
 /* routine.c */
 void	*routine(void *arg);
